move TestPatternParse into stra package && remove incorrect test file
